move breakpoints hook
